LayoutTests/fast/webgpu/regression/repro_290739-expected.txt

CONSOLE MESSAGE: promises created
CONSOLE MESSAGE: the end
CONSOLE MESSAGE: repro_290739.html
CONSOLE MESSAGE: Pass
